# docs/openapi.yaml
openapi: 3.1.0
info:
  title: API · Reservas Multi-sede (Peluquerías)
  version: "1.0"
  description: |
    API REST única que comparten la web pública, el bot de
    WhatsApp y el panel
    (doc 06). Horas en **UTC** (ISO 8601). Endpoints del panel autenticados con
    JWT (Bearer) y autorizados por rol. Errores con forma uniforme
    `{ "error": { "code", "message" } }`.
servers:
  - url: http://localhost:8000
    description: Desarrollo
...
security: []
paths:
...
  /api/v1/admin/staff:
    get:
      tags: [Panel]
      summary: Personal (por sede)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: location_id, in: query, schema: { type: integer } }]
      responses: { "200": { description: Personal } }
    post:
      tags: [Panel]
      summary: Alta de profesional
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string }, email: { type: [string, "null"] }, phone: { type: [string, "null"] }, location_ids: { type: array, items: { type: integer } }, service_ids: { type: array, items: { type: integer } } } } } } }
      responses: { "201": { description: Creado } }
  /api/v1/admin/staff/{id}:
    patch:
      tags: [Panel]
      summary: Editar profesional
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: Actualizado } }
  /api/v1/admin/staff/{id}/schedule:
    get:
      tags: [Panel]
      summary: Horario semanal del profesional
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: "Tramos (weekday 0=lun..6=dom)" } }
    post:
      tags: [Panel]
      summary: Reemplazar horario en una sede
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [location_id], properties: { location_id: { type: integer }, entries: { type: array, items: { type: object, properties: { weekday: { type: integer }, start_time: { type: string }, end_time: { type: string } } } } } } } } }
      responses: { "200": { description: Guardado } }
  /api/v1/admin/staff/{id}/calendar:
    get:
      tags: [Panel]
      summary: URL del feed iCal del profesional
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: "{ feed_url }" } }
  /api/v1/admin/staff/{id}/calendar/rotate:
    post:
      tags: [Panel]
      summary: Rotar el token del feed iCal
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { "200": { description: "{ feed_url } nuevo" } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...